Update msg.welcome
Just a few tweaks

I am honestly not really sure what belongs in an email like this. We already cover this in the man page, and in the new shell first screen, but I guess I like having it communicated in multiple places in multiple ways. Hopefully people catch at least one of them.