cron.daily/clean-lurkers: Proper user ranges
Previously, users between 1000 and 3000 were not concerned by
`clean-lurkers`.  Moreover, system users between 60000 and 65533 would
have been wrongly terminated.

The inclusion of the range 65535-4294967294 is for future-proofing,
since userdb and Debian explicitely support it.

Lastly, this fixed some potential issues with quoting,
and replaced a use of `awk` with `cut`.